# Speed up tests by shaving off subprocess when not needed