--- a/tests/test_cli.py
+++ b/tests/test_cli.py
@@ -26,6 +26,7 @@ from accelerate.commands.launch import _validate_launch_command, launch_command_
 from accelerate.test_utils import execute_subprocess_async
 from accelerate.test_utils.testing import (
     DEFAULT_LAUNCH_COMMAND,
+    capture_call_output,
     get_launch_command,
     path_in_accelerate_package,
     require_multi_device,
@@ -263,6 +264,7 @@ class ClusterConfigTester(unittest.TestCase):
         config = load_config_from_file(str(self.test_config_path / "0_30_0_sagemaker.yaml"))
 
 
+from accelerate.commands.tpu import tpu_command_launcher, tpu_command_parser
 class TpuConfigTester(unittest.TestCase):
     """
     Test case for verifying the `accelerate tpu-config` CLI passes the right `gcloud` command.
@@ -275,123 +277,116 @@ class TpuConfigTester(unittest.TestCase):
     base_output = "cd /usr/share"
     command_file = "tests/test_samples/test_command_file.sh"
     gcloud = "Running gcloud compute tpus tpu-vm ssh"
+    
+    def setUp(self):
+        self.parser = tpu_command_parser()
 
     def test_base(self):
-        output = run_command(
-            self.cmd
-            + ["--command", self.command, "--tpu_zone", self.tpu_zone, "--tpu_name", self.tpu_name, "--debug"],
-            return_stdout=True,
-        )
+        args = self.parser.parse_args(["--command", self.command, "--tpu_zone", self.tpu_zone, "--tpu_name", self.tpu_name, "--debug"])
+        output = capture_call_output(tpu_command_launcher, args)
         assert f"{self.gcloud} test-tpu --zone us-central1-a --command {self.base_output}; ls --worker all" in output
 
     def test_base_backward_compatibility(self):
-        output = run_command(
-            self.cmd
-            + [
-                "--config_file",
-                "tests/test_configs/0_12_0.yaml",
-                "--command",
-                self.command,
-                "--tpu_zone",
-                self.tpu_zone,
-                "--tpu_name",
-                self.tpu_name,
-                "--debug",
-            ],
-            return_stdout=True,
-        )
+        args = self.parser.parse_args(["--config_file", "tests/test_configs/0_12_0.yaml", "--command", self.command, "--tpu_zone", self.tpu_zone, "--tpu_name", self.tpu_name, "--debug"])
+        output = capture_call_output(tpu_command_launcher, args)
         assert f"{self.gcloud} test-tpu --zone us-central1-a --command {self.base_output}; ls --worker all" in output
 
     def test_with_config_file(self):
-        output = run_command(
-            self.cmd + ["--config_file", "tests/test_configs/latest.yaml", "--debug"], return_stdout=True
-        )
+        args = self.parser.parse_args(["--config_file", "tests/test_configs/latest.yaml", "--debug"])
+        output = capture_call_output(tpu_command_launcher, args)
         assert (
             f'{self.gcloud} test-tpu --zone us-central1-a --command {self.base_output}; echo "hello world"; echo "this is a second command" --worker all'
             in output
         )
 
     def test_with_config_file_and_command(self):
-        output = run_command(
-            self.cmd + ["--config_file", "tests/test_configs/latest.yaml", "--command", self.command, "--debug"],
-            return_stdout=True,
-        )
+        args = self.parser.parse_args(["--config_file", "tests/test_configs/latest.yaml", "--command", self.command, "--debug"])
+        output = capture_call_output(tpu_command_launcher, args)
+        # output = run_command(
+        #     self.cmd + ["--config_file", "tests/test_configs/latest.yaml", "--command", self.command, "--debug"],
+        #     return_stdout=True,
+        # )
         assert f"{self.gcloud} test-tpu --zone us-central1-a --command {self.base_output}; ls --worker all" in output
 
     def test_with_config_file_and_multiple_command(self):
-        output = run_command(
-            self.cmd
-            + [
-                "--config_file",
-                "tests/test_configs/latest.yaml",
-                "--command",
-                self.command,
-                "--command",
-                'echo "Hello World"',
-                "--debug",
-            ],
-            return_stdout=True,
-        )
+        args = self.parser.parse_args([
+            "--config_file", 
+            "tests/test_configs/latest.yaml", 
+            "--command", 
+            self.command, 
+            "--command", 
+            'echo "Hello World"',
+            "--debug"
+        ])
+        output = capture_call_output(tpu_command_launcher, args)
         assert (
             f'{self.gcloud} test-tpu --zone us-central1-a --command {self.base_output}; ls; echo "Hello World" --worker all'
             in output
         )
 
     def test_with_config_file_and_command_file(self):
-        output = run_command(
-            self.cmd
-            + ["--config_file", "tests/test_configs/latest.yaml", "--command_file", self.command_file, "--debug"],
-            return_stdout=True,
-        )
+        args = self.parser.parse_args(["--config_file", "tests/test_configs/latest.yaml", "--command_file", self.command_file, "--debug"])
+        output = capture_call_output(tpu_command_launcher, args)
+        # output = run_command(
+        #     self.cmd
+        #     + ["--config_file", "tests/test_configs/latest.yaml", "--command_file", self.command_file, "--debug"],
+        #     return_stdout=True,
+        # )
         assert (
             f'{self.gcloud} test-tpu --zone us-central1-a --command {self.base_output}; echo "hello world"; echo "this is a second command" --worker all'
             in output
         )
 
     def test_with_config_file_and_command_file_backward_compatibility(self):
-        output = run_command(
-            self.cmd
-            + [
-                "--config_file",
-                "tests/test_configs/0_12_0.yaml",
-                "--command_file",
-                self.command_file,
-                "--tpu_zone",
-                self.tpu_zone,
-                "--tpu_name",
-                self.tpu_name,
-                "--debug",
-            ],
-            return_stdout=True,
-        )
+        args = self.parser.parse_args(["--config_file", "tests/test_configs/0_12_0.yaml", "--command_file", self.command_file, "--tpu_zone", self.tpu_zone, "--tpu_name", self.tpu_name, "--debug"])
+        output = capture_call_output(tpu_command_launcher, args)
+        # output = run_command(
+        #     self.cmd
+        #     + [
+        #         "--config_file",
+        #         "tests/test_configs/0_12_0.yaml",
+        #         "--command_file",
+        #         self.command_file,
+        #         "--tpu_zone",
+        #         self.tpu_zone,
+        #         "--tpu_name",
+        #         self.tpu_name,
+        #         "--debug",
+        #     ],
+        #     return_stdout=True,
+        # )
         assert (
             f'{self.gcloud} test-tpu --zone us-central1-a --command {self.base_output}; echo "hello world"; echo "this is a second command" --worker all'
             in output
         )
 
     def test_accelerate_install(self):
-        output = run_command(
-            self.cmd + ["--config_file", "tests/test_configs/latest.yaml", "--install_accelerate", "--debug"],
-            return_stdout=True,
-        )
+        args = self.parser.parse_args(["--config_file", "tests/test_configs/latest.yaml", "--install_accelerate", "--debug"])
+        output = capture_call_output(tpu_command_launcher, args)
+        # output = run_command(
+        #     self.cmd + ["--config_file", "tests/test_configs/latest.yaml", "--install_accelerate", "--debug"],
+        #     return_stdout=True,
+        # )
         assert (
             f'{self.gcloud} test-tpu --zone us-central1-a --command {self.base_output}; pip install accelerate -U; echo "hello world"; echo "this is a second command" --worker all'
             in output
         )
 
     def test_accelerate_install_version(self):
-        output = run_command(
-            self.cmd
-            + [
-                "--config_file",
-                "tests/test_configs/latest.yaml",
-                "--install_accelerate",
-                "--accelerate_version",
-                "12.0.0",
-                "--debug",
-            ],
-            return_stdout=True,
-        )
+        args = self.parser.parse_args(["--config_file", "tests/test_configs/latest.yaml", "--install_accelerate", "--accelerate_version", "12.0.0", "--debug"])
+        output = capture_call_output(tpu_command_launcher, args)
+        # output = run_command(
+        #     self.cmd
+        #     + [
+        #         "--config_file",
+        #         "tests/test_configs/latest.yaml",
+        #         "--install_accelerate",
+        #         "--accelerate_version",
+        #         "12.0.0",
+        #         "--debug",
+        #     ],
+        #     return_stdout=True,
+        # )
         assert (
             f'{self.gcloud} test-tpu --zone us-central1-a --command {self.base_output}; pip install accelerate==12.0.0; echo "hello world"; echo "this is a second command" --worker all'
             in output