@muellerzr 2024-08-26
@BenjaminBossan yep exactly as it sounds, doing 2x subprocess is slower than 1x by a decent margin apparently.