@BenjaminBossan 2024-08-26
> yep exactly as it sounds, doing 2x subprocess is slower than 1x by a decent margin apparently

Hmm, you mean that `run_command` runs subprocess twice? How is that? Also, is it not surprising that this would lead to a 10x reduction?