@muellerzr 2024-08-26
Only when the underlying command calls subprocess (like the tpu ones)